requirements.txt
einops==0.8.0
ema_pytorch==0.5.1
imageio==2.35.1
lpips==0.1.4
matplotlib==3.9.2
moviepy==1.0.3
numpy==2.1.3
Pillow==11.0.0
pytorch_fid==0.3.0
scipy==1.14.1
tensorflow==2.18.0
timm==1.0.11
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
